move kinematic base to separate package
