+++ components/base/fake/fake.go
@@ -8,9 +8,8 @@ import (
 	"github.com/golang/geo/r3"
 
 	"go.viam.com/rdk/components/base"
-	"go.viam.com/rdk/referenceframe"
 	"go.viam.com/rdk/resource"
-	"go.viam.com/rdk/services/motion"
+	"go.viam.com/rdk/spatialmath"
 )
 
 func init() {
@@ -23,7 +22,7 @@ func init() {
 
 const (
 	defaultWidthMm               = 600
-	defaultMinimumTurningRadiusM = 0.8
+	defaultMinimumTurningRadiusM = 0
 )
 
 // Base is a fake base that returns what it was provided in each method.
@@ -31,14 +30,18 @@ type Base struct {
 	resource.Named
 	resource.TriviallyReconfigurable
 	CloseCount int
-	geometry   *referenceframe.LinkConfig
+	geometry   spatialmath.Geometry
 }
 
 // NewBase instantiates a new base of the fake model type.
-func NewBase(ctx context.Context, _ resource.Dependencies, conf resource.Config, _ golog.Logger) (base.Base, error) {
+func NewBase(_ context.Context, _ resource.Dependencies, conf resource.Config, _ golog.Logger) (base.Base, error) {
+	frame, err := conf.Frame.ParseConfig()
+	if err != nil {
+		return nil, err
+	}
 	return &Base{
 		Named:    conf.ResourceName().AsNamed(),
-		geometry: conf.Frame,
+		geometry: frame.Geometry(),
 	}, nil
 }
 
@@ -84,47 +87,4 @@ func (b *Base) Properties(ctx context.Context, extra map[string]interface{}) (ba
 		TurningRadiusMeters: defaultMinimumTurningRadiusM,
 		WidthMeters:         defaultWidthMm * 0.001, // convert to meters
 	}, nil
-}
-
-type kinematicBase struct {
-	*Base
-	model     referenceframe.Model
-	localizer motion.Localizer
-	inputs    []referenceframe.Input
-}
-
-// WrapWithKinematics creates a KinematicBase from the fake Base so that it satisfies the ModelFramer and InputEnabled interfaces.
-func (b *Base) WrapWithKinematics(
-	ctx context.Context,
-	localizer motion.Localizer,
-	limits []referenceframe.Limit,
-) (base.KinematicBase, error) {
-	geometry, err := base.CollisionGeometry(b.geometry)
-	if err != nil {
-		return nil, err
-	}
-	model, err := referenceframe.New2DMobileModelFrame(b.Name().ShortName(), limits, geometry)
-	if err != nil {
-		return nil, err
-	}
-	return &kinematicBase{
-		Base:      b,
-		model:     model,
-		localizer: localizer,
-		inputs:    make([]referenceframe.Input, len(model.DoF())),
-	}, nil
-}
-
-func (kb *kinematicBase) ModelFrame() referenceframe.Model {
-	return kb.model
-}
-
-func (kb *kinematicBase) CurrentInputs(ctx context.Context) ([]referenceframe.Input, error) {
-	return kb.inputs, nil
-}
-
-func (kb *kinematicBase) GoToInputs(ctx context.Context, inputs []referenceframe.Input) error {
-	_, err := kb.model.Transform(inputs)
-	kb.inputs = inputs
-	return err
 }
