+++ components/base/fake/fake_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/edaniels/golog"
 	"go.viam.com/test"
 
+	kinematicbase "go.viam.com/rdk/components/base/kinematicWrapper"
 	"go.viam.com/rdk/referenceframe"
 	"go.viam.com/rdk/resource"
 	"go.viam.com/rdk/services/motion"
@@ -37,7 +38,7 @@ func TestFakeBase(t *testing.T) {
 	localizer, err := motion.NewLocalizer(ctx, fakeSLAM)
 	test.That(t, err, test.ShouldBeNil)
 
-	kb, err := b.(*Base).WrapWithKinematics(ctx, localizer, limits)
+	kb, err := kinematicbase.WrapWithKinematics(ctx, b, localizer, limits)
 	test.That(t, err, test.ShouldBeNil)
 	expected := referenceframe.FloatsToInputs([]float64{10, 11, 0})
 	test.That(t, kb.GoToInputs(ctx, expected), test.ShouldBeNil)
