+++ services/navigation/builtin/builtin_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/edaniels/golog"
 	geo "github.com/kellydunn/golang-geo"
 	"go.viam.com/rdk/components/base"
-	_ "go.viam.com/rdk/components/base/fake"
 	fakebase "go.viam.com/rdk/components/base/fake"
+	kinematicbase "go.viam.com/rdk/components/base/kinematicWrapper"
 	_ "go.viam.com/rdk/components/movementsensor/fake"
 	"go.viam.com/rdk/config"
 	"go.viam.com/rdk/referenceframe"
@@ -105,7 +105,7 @@ func TestStartWaypoint(t *testing.T) {
 
 	localizer, err := motion.NewLocalizer(ctx, fakeslam.NewSLAM(slam.Named("foo"), logger))
 
-	kinematicBase, err := fakeBase.(base.KinematicWrappable).WrapWithKinematics(ctx, localizer, limits)
+	kinematicBase, err := kinematicbase.WrapWithKinematics(ctx, fakeBase, localizer, limits)
 	test.That(t, err, test.ShouldBeNil)
 
 	injectMovementSensor := inject.NewMovementSensor("test_movement")
